2nd refactoring
