+++ src/main/java/org/example/SplitFormula.java
@@ -1,0 +1,6 @@
+package org.example;
+
+interface SplitFormula {
+
+    String[] splitInputValue();
+}
